variscite: am62x_dram: fix dram_size overflow
Fixes R5 warning:
../common/am62x_dram.c:65:3: warning: conversion from ‘long long unsigned i
nt’ to ‘long unsigned int’ changes value from ‘4294967296’ to ‘0’ [-Woverflow]

+++ board/variscite/common/am62x_dram.c
@@ -52,7 +52,7 @@ int var_dram_init_mem_size_base(void) {
 	gd->ram_size = dram_size;
 
 	/* Set V2A_CTL_REG */
-	switch (dram_size) {
+	switch ((long long unsigned int) dram_size) {
 		case SZ_512M:
 			writel( ((readl(AM64_DDRSS_SS_BASE + 0x020) & ~0x3FF) | 0x1AD), AM64_DDRSS_SS_BASE + 0x020);
 			break;
